+++ CHANGELOG.md
@@ -22,6 +22,8 @@
 * Adjpositions are now stored in the database, and exported to XML ([#65])
 * `wn.Synset.ili` now returns an `ILI` object
 * `wn.remove()` now takes a `progess_handler` parameter
+* `wn.util.ProgressBar` uses a simpler formatting string with two new
+  computed variables
 
 ### Schema